WIP  - Separate js app into editorApp and frontendHighlighterApp apps  - Update enqueue-scripts.php to match the separated js  - Refactor scan process to use a checkPage module  - Remove full site scan code from checkPage module  - Remove auth header code from checkPage module
