+++ src/admin/sass/accessibility-checker-admin.scss
@@ -1,5 +1,5 @@
-@import "../../globals/sass/variables";
-@import "../../globals/sass/helpers";
+@import "../../common/sass/variables";
+@import "../../common/sass/helpers";
 
 .failed-text-color {
   color: $color-red;
