+++ src/editorApp/helpers.js
@@ -1,0 +1,23 @@
+import { settings } from './settings';
+
+
+export const info = (message) => {
+	if (settings.INFO_ENABLED) {
+		console.info(message);
+	}
+}
+
+
+export const debug = (message) => {
+	if (settings.DEBUG_ENABLED) {
+
+		if (location.href !== window.top.location.href) {
+			console.debug('DEBUG [ ' + location.href + ' ]');
+		}
+		if (typeof message !== 'object') {
+			console.debug('DEBUG: ' + message);
+		} else {
+			console.debug(message);
+		}
+	}
+}
