+++ src/frontendHighlighterApp/images/edac-emblem.png
