+++ src/frontendHighlighterApp/sass/app.scss
@@ -1,5 +1,5 @@
-@import "../../globals/sass/variables" ;
-@import "../../globals/sass/helpers";
+@import "../../common/sass/variables" ;
+@import "../../common/sass/helpers";
 
 
 body {
