+++ src/pageScanner/checks/always-fail.js
