+++ src/pageScanner/rules/color-contrast-failure.js
