+++ src/pageScanner/rules/custom-rule-1.js
